!!! FEATURE: Introduce commandcontext to access context variables
With this change a context class is introduced that
is forwarded to all commands that allows getting
the previously directly available context nodes
and now also the controller context.
With this future extendability is much easier as
the CommandInterface doesn’t need to change anymore
when new context variables are made available.

The feature is breaking as all custom commands need to be adapted.