BUGFIX: Prevent error in console from getCommands
This update introduces an additional endpoint to verify the user's access to available commands. Since this endpoint is not limited to the Admin or TerminalUser roles, we have implemented a safeguard to handle unsuccessful requests.

Conversely, there is also an extra request to retrieve commands when the user is authorized to manage them.

Fixes: #15